move initialize and to_json up to BasicModel
